+++ service/e2e_tests/fixtures/elasticsearch.py
@@ -1,0 +1,23 @@
+NL_MATERIAL = {
+    "title": "Didactiek van wiskundig denken",
+    "text": "Leermateriaal over wiskunde en didactiek op de universiteit.",
+    "url": "https://maken.wikiwijs.nl/91192/Wiskundedidactiek_en_ICT",
+    "description" : "Materiaal voor lerarenopleidingen en professionaliseringstrajecten gericht op wiskundedidactiek en ICT",
+    "language" : "nl",
+    "title_plain": "Wiskunde en Didactiek",
+    "text_plain": "Leermateriaal over wiskunde en didactiek op de universiteit.",
+    "external_id": "surf:oai:surfsharekit.nl:3522b79c-928c-4249-a7f7-d2bcb3077f10",
+    "copyright" : "cc-by-30",
+    "lom_educational_levels" : ["HBO"],
+    "publisher_date" : "2017-04-16T22:35:09+02:00",
+    "keywords": [],
+    "author": [
+        "BEGIN:VCARD\\nVERSION:3.0\\nPRODID:-//Sabre//Sabre VObject 3.3.5//EN\\nFN:Michel van Ast\\nN:van Ast;Michel;;;\\nUID:urn:wikiwijs:userid:2714696\\nEND:VCARD",
+        "BEGIN:VCARD\\nVERSION:3.0\\nPRODID:-//Sabre//Sabre VObject 3.3.5//EN\\nFN:Theo van den Bogaart\\nEND:VCARD",
+        "BEGIN:VCARD\\nVERSION:3.0\\nPRODID:-//Sabre//Sabre VObject 3.3.5//EN\\nFN:Marc de Graaf\\nEND:VCARD"
+    ],
+    "file_type" : "archive",
+    "disciplines": [],
+    "arrangement_collection_name" : "wikiwijsmaken"
+}
+
